// shared/generated/schemas/dtoTrackStatus.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Tune Backend API
 * Tune api for local consumers.
 * OpenAPI spec version: 1.0
 */

export type DtoTrackStatus = typeof DtoTrackStatus[keyof typeof DtoTrackStatus];


export const DtoTrackStatus = {
  TrackMissing: 'missing',
  TrackExist: 'exist',
  TrackQueued: 'queued',
} as const;