// shared/generated/schemas/sseSseDownloadEvent.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Tune Backend API
 * Tune api for local consumers.
 * OpenAPI spec version: 1.0
 */

export type SseSseDownloadEvent = typeof SseSseDownloadEvent[keyof typeof SseSseDownloadEvent];


export const SseSseDownloadEvent = {
  EventDownloadStarted: 'download-started',
  EventDownloadQueued: 'download-queued',
  EventDownloadFinished: 'download-finished',
  EventDownloadFailed: 'download-failed',
} as const;